Revert "cheri: malloc: avoid switch over uintptr_t"
This reverts commit dc8f9560104f0803e53fa23bfd011aa9878d5a97.

+++ malloc/mcheck-impl.c
@@ -74,7 +74,7 @@ checkhdr (const struct hdr *hdr)
        an error and report it.  */
     return MCHECK_OK;
 
-  switch (hdr->magic ^ ((unsigned long) hdr->prev + (unsigned long) hdr->next))
+  switch (hdr->magic ^ ((uintptr_t) hdr->prev + (uintptr_t) hdr->next))
     {
     default:
       status = MCHECK_HEAD;
